services/files/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
dependencies = [
  "core",
  "fastapi>=0.112.1",
  "pydantic-settings>=2.4.0",
  "fastapi-pagination>=0.12.26",
  "uvicorn>=0.30.6",
  "gunicorn>=23.0.0",
  "uvloop>=0.20.0",
  "SQLAlchemy>=2.0.29",
  "alembic>=1.13.2",
  "asyncpg>=0.29.0",
  "psycopg2-binary>=2.9.9",
  "async-fastapi-jwt-auth>=0.6.6",
  "orjson>=3.10.15",
  "passlib>=1.7.4",
  "redis>=5.2.1",
  "python-multipart>=0.0.20",
]
name = "files"
requires-python = ">=3.12"
version = "0.1.0"

[tool.hatch.build.targets.wheel]
packages = ["services/files"]

[tool.pytest.ini_options]
addopts = "-vv -p no:cacheprovider"
log_cli = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_level = "DEBUG"

[tool.uv.sources]
core = {workspace = true}
